Extract types into own modules from gmp protocols
